pipeline/requirements.txt
requests==2.32.3
feedparser==6.0.11
praw==7.7.1
nltk==3.9.1
langdetect==1.0.9
supabase==2.7.4
python-dotenv==1.0.1
